GSoC ideas: fix the USB idea
Isosynchronous transfers have been implemented in UHCI and EHCI since at
least 2011, and OHCI even before that. While it may not be completely
tested or have other problems, it is there and doesn't deserve a GSoC
idea.

Unfortunately one of our GSoC applicant thought it was fine to just copy
this idea as his proposal without discussing it with mentors first,
otherwise we could have noticed him of the problem.

At least it won't happen again next year.

Replace the isochronous USB idea with webcam nad USB-audio drivers,
which still need some work (and may allow to uncover bugs in isochronous
transfers in the underlying stack).