get-source: fix checkout and push instructions
On checkout, set up the commit hook to make sure commits get a change-id
automatically. This saves some confusion, and will be required in newer
versions of Gerrit.

On push, advise people to set a topic for their changes, which is very
useful to easily search for commits related to a specific topic (just
type it in Gerrit search field)