Show all the posts from a blog in its list (#464)
* Show all the posts from a blog in its list

Only show posts in that blog. That is, no articles from other sections
even if they have the same author as in blog/<blog>/_index.md.
Show all posts in that blog. That is, all the articles from blog/<blog>
even if they have a different author than the one in _index.md.

* Blog post: link to container blog