Add some scary warnings about makebootable
A common item in people having problem to boot Haiku is "so I tried to
run makebootable". In pretty much all cases, makebootable is not needed
and should not be used.

The only case where it is needed is when all of the following is true:
- Using BIOS (not EFI)
- Not using the anyboot image but a raw filesystem image (which we don't
  distribute as a standalone image anymore, you have to build it
  yourself or extract it from the anyboot)
- Not using Installer or jam @install to do the installation (both of
  these already take care of it)

Try to redirect people away from it as much as possible, since it is
quite unlikely that they will need such a crazy plan to install Haiku.