community/irc: Remove channes that are still on Freenode
I assume these are dead and there was just no one available to take care
of formally moving them. I think we should not keep any links with the
current incarnation of Freenode.

If you want to open a community channel for any of these languages (or
others), feel free to add it here. But pick a sane place to host it.