GSoC: some updates to the contribution page
Make sure students know they need to reach us on forums/mailing list/IRC
before submitting their application. It seems this is not currently
clear enough as some students applied only through the website and we
had to contact them to say them so.

It is a bit late for this year, but this page will be reused in the
future.