[GSoC 2024]: Implementing an Incremental Search Panel for Tracker - Final Report upload
This commit contains the Hugo Server upload for my Haiku Project. Although, I was not able to meet all of the expecations, I was able to set up a core edition of the project up and running. I am currently making some last minute changes after which those will be uploaded to gerrit as well :)